requirements.txt
pandas>=2.0.0
psycopg2-binary>=2.9.0
openai>=1.0.0
datasets>=2.0.0
jsonschema
python-dotenv